increase number size to i64
It's happy computing with `i128` but faster on `i64`.

20 primes starting at 10^18 takes about 50 sec (release optimized build) or 70 sec (debug build).

10^18: 50 sec
10^17: 14 sec
10^16: 6 sec

+++ src/main.rs
@@ -1,7 +1,7 @@
 // Finds prime numbers. A small project for learning Rust.
 // Maybe add a GUI or at least a CLI.
 
-type Num = i32;
+type Num = i64;
 
 // Return true if p is prime.
 // This implementation unrolls the loop once so it only has to test 2 potential
